fix web imports and extract hot reload handlers into seperate files
